+++ env.h
@@ -4,15 +4,16 @@
 # include <stdbool.h>
 
 // 環境変数とシェル変数をリストとして保持する
-typedef struct	s_var {
+typedef struct s_var {
 	const char			*key;
 	const char			*value;
 	bool				is_shell_var;
 	struct s_var		*next;
-}				t_var;
+}						t_var;
 
 void	free_vars(t_var *vars);
-t_var	*add_new_var(t_var **vars, const char *key, const char *value, bool is_shell_var);
+t_var	*add_new_var(t_var **vars,
+			const char *key, const char *value, bool is_shell_var);
 t_var	*kvstr2var(char *kvstr, bool is_shell_var);
 t_var	*environ2vars(char **environ);
 char	**vars2environ(t_var *vars);
@@ -24,7 +25,7 @@ char	**get_colon_units(const char *str, const char *default_str);
 char	*get_val_from_kvstr(const char *kvstr, char delimiter);
 char	*expand_env_var(char *str);
 char	*generate_kvstr(const char *key, const char *value);
-int	ft_setenv(const char *key, const char *value, bool is_shell_var);
+int		ft_setenv(const char *key, const char *value, bool is_shell_var);
 int		ft_unsetenv(const char *key);
 
 // コマンドの終了ステータス
